add local database

+++ MyApp/data/src/main/java/anna/poddubnaya/data/repository/UserRepositoryImpl.java
@@ -1,6 +1,8 @@
 package anna.poddubnaya.data.repository;
 
 import android.content.Context;
+import android.net.ConnectivityManager;
+import android.net.NetworkInfo;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,10 +12,10 @@ import anna.poddubnaya.data.database.UserDao;
 import anna.poddubnaya.data.entity.User;
 import anna.poddubnaya.data.rest.RestService;
 import anna.poddubnaya.domain.entity.UserEntity;
-import anna.poddubnaya.domain.usecases.UserRepository;
+import anna.poddubnaya.domain.repository.UserRepository;
 import io.reactivex.Completable;
+import io.reactivex.Flowable;
 import io.reactivex.Observable;
-import io.reactivex.ObservableSource;
 import io.reactivex.functions.Consumer;
 import io.reactivex.functions.Function;
 
@@ -23,6 +25,7 @@ public class UserRepositoryImpl implements UserRepository {
     private Context context;
     private RestService restService;
     private UserDao userDao;
+    private boolean connected = false;
 
     public UserRepositoryImpl(Context context, RestService restService, AppDatabase database) {
         this.context = context;
@@ -31,9 +34,20 @@ public class UserRepositoryImpl implements UserRepository {
     }
 
     @Override
-    public Observable<UserEntity> get(String id) {
-        return restService
-                .loadUserById(id)
+    public Flowable<UserEntity> get(String id) {
+        Flowable<User> user;
+        if (checkNetwork()) {
+            user = restService.loadUserById(id);
+        } else {
+            user = userDao.getById(id).take(1).map(new Function<List<User>, User>() {
+                @Override
+                public User apply(List<User> users) throws Exception {
+                    return users.get(0);
+                }
+            });
+
+        }
+        return user
                 .map(new Function<User, UserEntity>() {
                     @Override
                     public UserEntity apply(User user) throws Exception {
@@ -45,27 +59,23 @@ public class UserRepositoryImpl implements UserRepository {
                 });
     }
 
+
     @Override
-    public Observable<List<UserEntity>> get() {
-
-        return restService
-                .loadUsers()
-                .doOnNext(new Consumer<List<User>>() {
-                    @Override
-                    public void accept(List<User> userList) throws Exception {
-                        userDao.deleteAll();
-                        userDao.insert(userList);
-                    }
-                })
-                .onErrorResumeNext(new Function<Throwable, ObservableSource<? extends List<User>>>() {
-                    @Override
-                    public ObservableSource<? extends List<User>> apply(Throwable throwable) throws Exception {
-
-                        //желательно проверить ошибку - если интернет - то берем из базы данных
-                        //если не интернет - то надо выкинуть саму ошибку
-                        return userDao.getAll().toObservable().take(1);
-                    }
-                })
+    public Flowable<List<UserEntity>> get() {
+        Flowable<List<User>> users;
+        if (checkNetwork()) {
+            users = restService.loadUsers()
+                    .doOnNext(new Consumer<List<User>>() {
+                        @Override
+                        public void accept(List<User> userList) throws Exception {
+                            userDao.deleteAll();
+                            userDao.insert(userList);
+                        }
+                    });
+        } else {
+            users = userDao.getAll().take(1);
+        }
+        return users
                 .map(new Function<List<User>, List<UserEntity>>() {
                     @Override
                     public List<UserEntity> apply(List<User> users) throws Exception {
@@ -80,21 +90,43 @@ public class UserRepositoryImpl implements UserRepository {
                         }
                         return list;
                     }
-                }) ;
+                });
     }
 
     @Override
     public Completable save(UserEntity userEntity) {
-        User user = new User();
-        user.setUsername(userEntity.getUsername());
-        user.setAge(userEntity.getAge());
-        user.setObjectId(userEntity.getObjectId());
-        user.setProfileUrl(userEntity.getProfileUrl());
-        return restService.save(user);
+        if (checkNetwork()) {
+            User user = new User();
+            user.setUsername(userEntity.getUsername());
+            user.setAge(userEntity.getAge());
+            user.setObjectId(userEntity.getObjectId());
+            user.setProfileUrl(userEntity.getProfileUrl());
+            return restService.save(user);
+        } else {
+            return Completable.error(new Throwable("no internet connection"));
+        }
+
     }
 
     @Override
     public Completable remove(String id) {
-        return restService.remove(id);
+        if (checkNetwork()) {
+            return restService.remove(id);
+        } else {
+            return Completable.error(new Throwable("no internet connection"));
+        }
+    }
+
+
+    public boolean checkNetwork() {
+        boolean isConnected = false;
+        ConnectivityManager connectivityManager = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
+        if (connectivityManager != null) {
+            if (connectivityManager.getNetworkInfo(ConnectivityManager.TYPE_MOBILE).getState() == NetworkInfo.State.CONNECTED ||
+                    connectivityManager.getNetworkInfo(ConnectivityManager.TYPE_WIFI).getState() == NetworkInfo.State.CONNECTED) {
+                isConnected = true;
+            }
+        }
+        return isConnected;
     }
 }
